feat: pipeline false alarm checks, 64-bit chain counts, Markov NTLM8/9 tests, improved coverage model
- Pipeline false alarm checking in crackalack_lookup (launch/harvest pattern for GPU overlap)
- Widen chain count fields to uint64_t to support large tables (>4B chains)
- Add per-kernel GWS tuning (get_optimal_gws now takes kernel name)
- Improve coverage model in crackalack_plan with iterative merge accounting
- Add Markov NTLM8/NTLM9 chain test files and unit test registrations
- Update test_markov_accuracy.sh with train/test split and improved params
- Fix Metal kernel address space qualifiers and buffer argument handling
- Various kernel fixes for Markov fast-path (pspace_table, sorted_bigram)

+++ parallel_sort.h
@@ -13,6 +13,6 @@
  * Returns 0 on success, -1 on allocation failure (caller should fall back
  * to single-threaded qsort). Falls back internally to single-threaded
  * qsort when num_chains < 1024 or num_threads <= 1. */
-int parallel_sort_rt(uint64_t *data, unsigned int num_chains, int num_threads);
+int parallel_sort_rt(uint64_t *data, uint64_t num_chains, int num_threads);
 
 #endif /* PARALLEL_SORT_H */
